requirements.txt
ddt==1.2.1
dj-database-url==0.5.0
Django==2.2.8
django-crispy-forms==1.8.0
django-heroku==0.3.1
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.5
gunicorn==20.0.0
isort==4.3.21
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
pytest==5.3.1
pytest-cov==2.8.1
pytest-django==3.7.0
pytz==2019.3
snowballstemmer==2.0.0
sqlparse==0.3.0
tox==3.14.1
whitenoise==4.1.4